Unify FreeC and Pull
